#6 Add integration tests
