+++ registry/registry_int.go
